+++ priv/repo/migrations/20171104012312_create_clips.exs
@@ -4,7 +4,7 @@ defmodule Livedub.Repo.Migrations.CreateClips do
   def change do
     create table(:clips) do
       add :url, :string, null: false
-      add :start_time, :bigint, null: false
+      add :start_time, :float, null: false
       add :user_id, references(:users, on_delete: :nothing), null: false
       add :jam_id, references(:jams, on_delete: :nothing), null: false
